Create 1768-MergeStringsAlternatively.cpp
Well we need a string whose length is equal to the combination of word1 and word2, then char by char, we can traverse the words one by one, if we don't exceed the length of the word, we add the current character to the merged word and move one place forward in the merged word, otherwise we skip this word and move to next. This entire thing goes on till we reach the end of the merged word. Yep that's it !