Create FlowerPot.cpp
Intuition
Well to put a flower such that it has no adjacent flower, you have to check the position itself, the position before it and also the position after it. If all is safe, you can plant a flower there and just then increase the count of flowers.

Approach
There are 2 things to remember, the first 2 places and the last 2 places, because when we check in groups of 3, those 2 places can be left unchecked or be unreachable. So we check the first condition, then start from the first index and go till the second last index and check in groups of 3 and then finally check the last 2 places.
Also we need to put special edge conditions for arrays which have only 1 or 2 places. Yah, that's all ig..

Complexity
Time complexity:

The loop runs in O(sz) time, where sz is the size of the flowerbed.
So we can say it has O(n) time complexity.
Space complexity:

The additional space used by integers (j, sz, i) and other variables is constant.
So we can say it has O(1) space complexity